requirements.txt
click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe==
speedtest-cli
Werkzeug
